GETP-216 refactor: storage 컴포넌트의 infrastructure 모듈과 관련된 예외들이 FileStorageException을 상속하도록 변경
